# swagger/swagger.json
{
  "swagger":"2.0",
  "info":{
    "description":"Api Documentation",
    "version":"1.0",
    "title":"Api Documentation",
    "termsOfService":"urn:tos",
    "contact":{

    },
    "license":{
      "name":"Apache 2.0",
      "url":"http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host":"localhost:8080",
  "basePath":"/",
  "tags":[
    {
      "name":"custom-error-controller",
      "description":"Custom Error Controller"
    },
    {
      "name":"profile-controller",
      "description":"Profile Controller"
    },
    {
      "name":"widget-controller",
      "description":"Widget Controller"
    }
  ],
  "paths":{
    "/api/v1/widgets":{
      "get":{
        "tags":[
          "widget-controller"
        ],
        "summary":"getAll",
        "operationId":"getAllUsingGET",
        "produces":[
          "*/*"
        ],
        "parameters":[
          {
            "name":"limit",
            "in":"query",
            "description":"limit",
            "required":false,
            "type":"integer",
            "format":"int32"
          },
          {
            "name":"offset",
            "in":"query",
            "description":"offset",
            "required":false,
            "type":"integer",
            "format":"int32"
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "type":"array",
              "items":{
                "$ref":"#/definitions/Widget"
              }
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "post":{
        "tags":[
          "widget-controller"
        ],
        "summary":"create",
        "operationId":"createUsingPOST",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "*/*"
        ],
        "parameters":[
          {
            "in":"body",
            "name":"widget",
            "description":"widget",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Widget"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/Widget"
            }
          },
          "201":{
            "description":"Created"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "put":{
        "tags":[
          "widget-controller"
        ],
        "summary":"update",
        "operationId":"updateUsingPUT",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "*/*"
        ],
        "parameters":[
          {
            "in":"body",
            "name":"widget",
            "description":"widget",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Widget"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/Widget"
            }
          },
          "201":{
            "description":"Created"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      }
    },
    "/api/v1/widgets/{id}":{
      "get":{
        "tags":[
          "widget-controller"
        ],
        "summary":"getById",
        "operationId":"getByIdUsingGET",
        "produces":[
          "*/*"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/Widget"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "delete":{
        "tags":[
          "widget-controller"
        ],
        "summary":"delete",
        "operationId":"deleteUsingDELETE",
        "produces":[
          "*/*"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"path",
            "description":"id",
            "required":true,
            "type":"integer",
            "format":"int64"
          }
        ],
        "responses":{
          "200":{
            "description":"OK"
          },
          "204":{
            "description":"No Content"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          }
        }
      }
    },
    "/error":{
      "get":{
        "tags":[
          "custom-error-controller"
        ],
        "summary":"error",
        "operationId":"errorUsingGET",
        "produces":[
          "*/*"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "type":"object",
              "additionalProperties":{
                "type":"object"
              }
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "head":{
        "tags":[
          "custom-error-controller"
        ],
        "summary":"error",
        "operationId":"errorUsingHEAD",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "*/*"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "type":"object",
              "additionalProperties":{
                "type":"object"
              }
            }
          },
          "204":{
            "description":"No Content"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          }
        }
      },
      "post":{
        "tags":[
          "custom-error-controller"
        ],
        "summary":"error",
        "operationId":"errorUsingPOST",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "*/*"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "type":"object",
              "additionalProperties":{
                "type":"object"
              }
            }
          },
          "201":{
            "description":"Created"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "put":{
        "tags":[
          "custom-error-controller"
        ],
        "summary":"error",
        "operationId":"errorUsingPUT",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "*/*"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "type":"object",
              "additionalProperties":{
                "type":"object"
              }
            }
          },
          "201":{
            "description":"Created"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "delete":{
        "tags":[
          "custom-error-controller"
        ],
        "summary":"error",
        "operationId":"errorUsingDELETE",
        "produces":[
          "*/*"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "type":"object",
              "additionalProperties":{
                "type":"object"
              }
            }
          },
          "204":{
            "description":"No Content"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          }
        }
      },
      "options":{
        "tags":[
          "custom-error-controller"
        ],
        "summary":"error",
        "operationId":"errorUsingOPTIONS",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "*/*"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "type":"object",
              "additionalProperties":{
                "type":"object"
              }
            }
          },
          "204":{
            "description":"No Content"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          }
        }
      },
      "patch":{
        "tags":[
          "custom-error-controller"
        ],
        "summary":"error",
        "operationId":"errorUsingPATCH",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "*/*"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "type":"object",
              "additionalProperties":{
                "type":"object"
              }
            }
          },
          "204":{
            "description":"No Content"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          }
        }
      }
    },
    "/profile":{
      "get":{
        "tags":[
          "profile-controller"
        ],
        "summary":"listAllFormsOfMetadata",
        "operationId":"listAllFormsOfMetadataUsingGET",
        "produces":[
          "*/*"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "$ref":"#/definitions/RepresentationModel«object»"
            }
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          },
          "404":{
            "description":"Not Found"
          }
        }
      },
      "options":{
        "tags":[
          "profile-controller"
        ],
        "summary":"profileOptions",
        "operationId":"profileOptionsUsingOPTIONS",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "*/*"
        ],
        "responses":{
          "200":{
            "description":"OK",
            "schema":{
              "type":"object"
            }
          },
          "204":{
            "description":"No Content"
          },
          "401":{
            "description":"Unauthorized"
          },
          "403":{
            "description":"Forbidden"
          }
        }
      }
    }
  },
  "definitions":{
    "Links":{
      "type":"object",
      "properties":{
        "empty":{
          "type":"boolean"
        }
      },
      "title":"Links"
    },
    "RepresentationModel«object»":{
      "type":"object",
      "properties":{
        "links":{
          "$ref":"#/definitions/Links"
        }
      },
      "title":"RepresentationModel«object»"
    },
    "Widget":{
      "type":"object",
      "properties":{
        "height":{
          "type":"integer",
          "format":"int32"
        },
        "id":{
          "type":"integer",
          "format":"int64"
        },
        "modifiedAt":{
          "type":"string",
          "format":"date-time"
        },
        "width":{
          "type":"integer",
          "format":"int32"
        },
        "x":{
          "type":"integer",
          "format":"int32"
        },
        "y":{
          "type":"integer",
          "format":"int32"
        },
        "z":{
          "type":"integer",
          "format":"int32"
        }
      },
      "title":"Widget"
    }
  }
}